+++ diive/core/plotting/__init__.py
@@ -1,2 +1,2 @@
-from . import styles
-from . import plotfuncs
+# from . import styles
+# from . import plotfuncs