+++ diive/pkgs/__init__.py
@@ -1,5 +1,5 @@
-from diive.pkgs import analyses
-from diive.pkgs import corrections
-from diive.pkgs import createvar
-from diive.pkgs import flux
-from diive.pkgs import qaqc
+# from diive.pkgs import analyses
+# from diive.pkgs import corrections
+# from diive.pkgs import createvar
+# from diive.pkgs import flux
+# from diive.pkgs import qaqc