+++ diive/pkgs/flux/__init__.py
@@ -1,2 +1,1 @@
-import diive.pkgs.flux.criticaldays
-import diive.pkgs.flux.co2_penalty
+