+++ CHANGELOG.md
@@ -2,22 +2,33 @@
 
 ![DIIVE](images/logo_diive1_256px.png)
 
+## v0.84.1 | 8 Nov 2024
+
+### Bugfixes
+
+- Removed invalid imports
+
+### Tests
+
+- Added test case for `diive` imports (`tests.test_imports.TestImports.test_imports`)
+- 52/52 unittests ran successfully
+
 ## v0.84.0 | 7 Nov 2024
 
-# New features
+### New features
 
 - New class `BinFitterCP` for fitting function to binned data, includes confidence interval and prediction interval (
   `diive.pkgs.fits.fitter.BinFitterCP`)
 
 ![DIIVE](images/BinFitterCP_diive_v0.84.0.png)
 
-## Additions
+### Additions
 
 - Added small function to detect duplicate entries in lists (`diive.core.funcs.funcs.find_duplicates_in_list`)
 - Added new filetype (`diive/configs/filetypes/ETH-MERCURY-CSV-20HZ.yml`)
 - Added new filetype (`diive/configs/filetypes/GENERIC-CSV-HEADER-1ROW-TS-END-FULL-NS-20HZ.yml`)
 
-## Bugfixes
+### Bugfixes
 
 - Not directly a bug fix, but when reading EddyPro fluxnet files with `LoadEddyProOutputFiles` (e.g., in the flux
   processing chain) duplicate columns are now automatically renamed by adding a numbered suffix. For example, if two